Move internal package to be within executor, move stats to internal package
