+++ simulation/executor/internal/stats/schema.sql
